# Performance improvement of `cupy.in1d`