Rel. #4018

master:
```
in1d                :    CPU: 8473.098 us   +/-52.090 (min: 8446.542 / max: 8763.044) us     GPU-0:120539.443 us   +/-47.472 (min:120487.938 / max:120788.994) us
```

This PR:
```
in1d                :    CPU: 8428.611 us   +/-44.928 (min: 8408.114 / max: 8835.957) us     GPU-0:64765.102 us   +/-43.197 (min:64736.259 / max:65145.859) us
```

Benchmark:
```py
import cupy
from cupy import testing
from cupyx.profiler import benchmark


shape = (64, 1024, 1024)
x1 = testing.shaped_random(shape, xp=cupy, dtype='l', seed=0, scale=100000000)
x2 = testing.shaped_random(shape, xp=cupy, dtype='l', seed=1, scale=100000000)
perf = benchmark(cupy.in1d, (x1, x2), n_repeat=100)

print(perf)
print(cupy.in1d(x1, x2).sum())
```

cc: @mrkwjc